+++ src/qcommon/qcommon.h
@@ -696,7 +696,7 @@ void  Com_Error( int code, const char *fmt, ... );
 void        Com_Quit_f( void );
 
 int         Com_Milliseconds( void );   // will be journaled properly
-unsigned    Com_BlockChecksum( const void *buffer, int length );
+unsigned    int Com_BlockChecksum( const void *buffer, size_t length );
 unsigned    Com_BlockChecksumKey( void *buffer, int length, int key );
 int         Com_HashKey( char *string, int maxlen );
 int         Com_Filter( char *filter, char *name, int casesensitive );
